refactor: update coordinates to latitude/longitude, and upgrade financial precision

+++ backend/src/main/java/com/autoride/dto/request/UpdateBalanceRequest.java
@@ -3,8 +3,10 @@ package com.autoride.dto.request;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record UpdateBalanceRequest(
         @NotNull(message = "Amount is required")
         @Positive(message = "Amount must be positive")
-        Double amount
+        BigDecimal amount
 ) {}
